rename durationToClockDuration in videoDurationToTime
also move the function to the TournesolVideoCard as for now only
the video cards require it. it avoids creating a huge utils file
